refactor: move api and entities to api dir
